Add files via upload
Tools to automate daily review of stock trades.  This is an essential but time consuming task. The intentions is to let the user keep focus on what was learned today and not on the mechanics of book keeping. The first version requires the  CVS export file from DAS Trader Pro Trades Window. Future versions be more general.